project/requirements.txt
streamlit
pathlib
shutil
pylovepdf
pillow